src/requirements.txt
APScheduler==3.6.3
async-generator==1.10
attrs==22.1.0
cachetools==4.2.2
certifi==2022.6.15
charset-normalizer==2.1.0
h11==0.13.0
idna==3.3
outcome==1.2.0
PySocks==1.7.1
python-telegram-bot==13.13
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
selenium==4.4.3
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
tornado==6.2
trio==0.21.0
trio-websocket==0.9.2
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.10
wsproto==1.2.0
